Refactor url encoder api
